新增新闻投票系统缓存失效处理类，增加新闻数据统计

+++ mlxc-news/mlxc-news-api/src/main/java/com/jxxy/mlxc/news/api/dto/NewsDto.java
@@ -42,7 +42,7 @@ public class NewsDto extends BaseDto {
 	/**
 	 * 点赞数量
 	 */
-	private Long grade;
+	private Integer grade;
 	/**新闻或文章*/
 	private Integer type;
 	/**吃住行玩*/
